♻️ Refactor: strongly-type AuthErrors
